fix(apps/ekiosk): resolve multiple build warnings
- PluginService.cpp: Convert SM_SHUTTINGDOWN macro to constexpr constant
- TerminalService.cpp: Replace do-while with for loop, fix narrowing conversions
- TerminalService.cpp: Use braced initializer list for return statement
- CashDispenserManager.cpp: Change math.h to cmath, fix narrowing conversion
- CryptService.cpp: Add NOLINT for intentional assignment-in-if patterns

+++ apps/EKiosk/src/Services/CashDispenserManager.cpp
@@ -13,7 +13,7 @@
 #include <SDK/PaymentProcessor/Settings/TerminalSettings.h>
 
 #include <algorithm>
-#include <math.h>
+#include <cmath>
 #include <numeric>
 
 #include "DatabaseUtils/IHardwareDatabaseUtils.h"
@@ -300,7 +300,7 @@ CashDispenserManager::getItem_DataSet(PPSDK::TPaymentAmount aAmount) {
         if (!m_FailedDispensers.contains(dispenser)) {
             auto cashList = m_CurrencyCashList[m_Dispensers[dispenser]];
 
-            for (int i = 0; i < cashList.size(); i++) {
+            for (int i = 0; i < static_cast<int>(cashList.size()); i++) {
                 if (dispenser->isDeviceReady(i) && (cashList[i].count != 0) &&
                     (cashList[i].nominal != 0) && (cashList[i].nominal <= aAmount)) {
                     int count = (cashList[i].count > 0) ? cashList[i].count : 0;
